cleaned up campaign controller, added validations to create campaign form

+++ app/controllers/campaigns_controller.rb
@@ -1,4 +1,7 @@
 class CampaignsController < ApplicationController
+  include CreateFormValidation
+  include CampaignHelper
+
   before_filter :auth_user, :except => [:index, :show]
   before_filter :check_same_user, :only => [:edit, :update, :destroy]
 
@@ -24,15 +27,8 @@ class CampaignsController < ApplicationController
 
   def show
     session.delete(:campaign_id) if session[:campaign_id]
-    @support = current_user.supported_campaigns if current_user
     @campaign = Campaign.find_by_id(params[:id])
-    @user_supported = CampaignUser.where(user_id: current_user.id, campaign_id: @campaign.id, user_type: SUPPORTER) if current_user
-    support_ids = CampaignUser.where(campaign_id: @campaign.id, user_type: SUPPORTER).pluck("user_id")
-    donor_ids = CampaignUser.where(campaign_id: @campaign.id, user_type: DONOR).pluck("user_id")
-
-    @supporters = support_ids.map { |id| User.find(id)}
-    @donors = donor_ids.map { |id| User.find(id) }
-    @video = UrlToMediaTag.convert(@campaign.video_url, width: 540, height: 320)
+    @campaign_show = CampaignHelper.render_objects_for_view(current_user, @campaign)
   end
 
   def edit
@@ -54,32 +50,43 @@ class CampaignsController < ApplicationController
 
   def filter_topic
     topic = Topic.find_by_title(params[:topic])
-    p campaigns = Campaign.where(topic_id: topic.id)
+    campaigns = Campaign.where(topic_id: topic.id)
     render json: render_to_string(partial: 'filtered_opp_campaigns', locals: { camp: campaigns }).to_json
   end
 
   def editable_form
-    @campaign = Campaign.find_by_id(session[:campaign_build])
-    @campaign.update_attribute('pitch', params[:form])
-    respond_to do |format|
-      if @campaign.save
-        format.js { render partial: 'campaigns/editable_form' }
-      else
-        format.json { render json: @campaign.errors.full_messages, status: :unprocessable_entity }
+    if params[:form].length < 520
+      errors = "Pitch must be a minimum length of 520 characters"
+      render json: errors.to_json 
+    else
+      @campaign = Campaign.find_by_id(session[:campaign_build])
+      @campaign.update_attribute('pitch', params[:form])
+      respond_to do |format|
+        if @campaign.save
+          format.js { render partial: 'campaigns/editable_form' }
+        else
+          format.json { render json: @campaign.errors.full_messages, status: :unprocessable_entity }
+        end
       end
     end
   end
   
   def finalize_campaign
-    @campaign = Campaign.find_by_id(session[:campaign_build])
-    @campaign.update_attributes(params[:campaign])
-    if @campaign.save
-      ScheduledWorker.perform_at(@campaign.funding_deadline, @campaign.id)
-      # UserMailer.campaign_new_email(current_user, @campaign).deliver
-      session.delete(:campaign_build)
-      render nothing: true
+    errors = CreateFormValidation.form_three_validations(params[:campaign])
+    if errors.count == 1
+      @campaign = Campaign.find_by_id(session[:campaign_build])
+      @campaign.update_attributes(params[:campaign])
+      @campaign.update_attribute(:status, ACTIVE)
+      if @campaign.save
+        ScheduledWorker.perform_at(@campaign.funding_deadline, @campaign.id)
+        # UserMailer.campaign_new_email(current_user, @campaign).deliver
+        session.delete(:campaign_build)
+        render nothing: true
+      else
+        render json: {:error => @campaign.errors.full_messages}.to_json, :status => :unprocessable_entity
+      end
     else
-      render json: {:error => @campaign.errors.full_messages}.to_json, :status => :unprocessable_entity
+      render json: errors.to_json
     end
   end
 
@@ -112,3 +119,4 @@ class CampaignsController < ApplicationController
       end
     end
 end
+
